requirements.txt
# German Pronunciation Diagnostic App (L2-Trainer) - Requirements

# Core ML and Audio Processing
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.24.0
scipy>=1.10.0

# G2P (Grapheme-to-Phoneme)
phonemizer>=3.0.0

# Alignment
biopython>=1.81

# VAD (Voice Activity Detection)
silero-vad>=4.0.0
# Alternative: webrtcvad (optional, for fallback)

# Frontend
gradio>=4.0.0

# Utilities
pathlib2>=2.3.7; python_version < '3.4'


